requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
anyio==3.3.0
appdirs==1.4.4
APScheduler==3.7.0
arrow==1.1.1
asgiref==3.4.1
async-timeout==4.0.2
Babel==2.9.1
binaryornot==0.4.4
black==21.7b0
blinker==1.4
bypy==1.7.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cn2an==0.5.11
colorama==0.4.4
cookiecutter==1.7.3
dill==0.3.4
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.18.2
humanreadable==0.1.0
hypercorn==0.13.2
hyperframe==5.2.0
idna==3.2
iso8601==0.1.16
itsdangerous==2.0.1
Jinja2==3.0.1
jinja2-time==0.2.0
loguru==0.5.3
MarkupSafe==2.0.1
mbstrdecoder==1.0.1
mpmath==1.2.1
msgpack==1.0.2
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nb-cli==0.5.1
neovim==0.3.1
nonebot-adapter-cqhttp==2.0.0a15
nonebot-plugin-apscheduler==0.1.2
nonebot2==2.0.0a16
pathspec==0.9.0
poyo==0.5.0
priority==2.0.0
prompt-toolkit==3.0.19
pyaes==1.6.1
pyasn1==0.4.8
pydantic==1.8.2
pyfiglet==0.8.post1
pyglet==1.5.19
pygtrie==2.4.2
pynvim==0.4.3
pypika-tortoise==0.1.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
python-socks==2.0.1
pytz==2021.1
PyYAML==5.4.1
quart==0.16.2
regex==2021.8.3
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
sympy==1.8
Telethon==1.24.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tomlkit==0.7.2
tortoise-orm==0.18.0
typepy==1.2.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.2
wsproto==1.0.0
yapf==0.31.0
